Move send signal/sequence into terminalContrib
